Preparing to handle high res textures. As locations are small numbers and int. But will be needing to scale up. Vectors can be multiplied together to scale so 1/4 the lines of code to do the tweaks I think.  I'll do changes in a bit at a time I hope. :)

+++ FF8/EntryGroup.cs
@@ -79,8 +79,8 @@ namespace FF8
                             Width = list[0].Width,
                             Height = list[0].Height,
                             Tile = Vector2.UnitX,
-                            Offset = new Point(list[0].Width,0),
-                            End = new Point(-list[0].Width,0)
+                            Offset = new Point((int)list[0].Width,0),
+                            End = new Point((int)-list[0].Width,0)
                         });
                 }
                 list.Add(entry);
